Implemented tabs which can be registered to the preview. Tabs can be switched for different previews. Added a get parameter "debug" to set the debugmode on or off from the stage
