+++ preview/src/com/badlogicgames/gdxskins/client/skins/info/SkinContainer.java
@@ -1,0 +1,25 @@
+package com.badlogicgames.gdxskins.client.skins.info;
+
+import com.badlogic.gdx.scenes.scene2d.ui.Skin;
+
+/**
+ * Created by Lyze on 01.06.2016.
+ */
+public class SkinContainer {
+    public SkinMeta skinMeta;
+    public Skin skin;
+
+    public SkinContainer (SkinMeta meta, Skin skin) {
+        this.skinMeta = meta;
+        this.skin = skin;
+    }
+
+    public void dispose () {
+        skin.dispose();
+    }
+
+    @Override
+    public String toString () {
+        return skinMeta.title;
+    }
+}
