+++ preview/src/gdxskins.gwt.xml
@@ -2,5 +2,5 @@
 <!DOCTYPE module PUBLIC "-//Google Inc.//DTD Google Web Toolkit trunk//EN" "http://google-web-toolkit.googlecode.com/svn/trunk/distro-source/core/src/gwt-module.dtd">
 <module>
     <source path="com/badlogicgames/gdxskins" />
-    <extend-configuration-property name="gdx.reflect.include" value="com.badlogicgames.gdxskins.client.SkinMeta"/>
+    <extend-configuration-property name="gdx.reflect.include" value="com.badlogicgames.gdxskins.client.skins.info.SkinMeta"/>
 </module>
